% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portland_crashes_2018.R
\docType{data}
\name{portland_crashes_2018}
\alias{portland_crashes_2018}
\title{2018 Portland Car Crashes}
\format{A dataframe with 3940 rows and 40 variables:

\describe{
  \item{crash_id}{Unique identification number assigned to each crash}
  \item{partic_id}{Unique identification numer assigned to each person involved in a crash}
  \item{vhcl_id}{Unique identification number assigned to each vehicle involved in a crash}
  \item{partic_no}{Identifier for participants within a specific crash e.g., for a 3 person crash, person 1 is 1, person 2 is 2, person 3 is 3}
  \item{vhcl_no}{Identifier for vehicles inolved in a specific crash e.g., for a 3 vehicle crash, vehicle 1 is 1, vehicle 2 is 2, vehicle 3 is 3}
  \item{partic_type}{Participants are classified as drvr (driver), psng (passenger), or prkd (occupant of a parked vehicle). Pedestrians are not included in the data set because they are not associated with a vehicle ID (vhcl_id)}
  \item{sex}{Sex of participant}
  \item{age}{Age of participant}
  \item{drvr_lic_stat}{Driver's license status of drivers of vehicles involved in a crash were recorded as OR (valid Oregon license or permit), unknown (information not available), suspended, none (no license), and other state (valid license from another state)}
  \item{inj_svrty}{Severity of injury suffered by crash participant. Recorded as kill (fatal), none (no injury), inja (suspected serious injury), injb (suspected minor injury), injc (possible injury)}
  \item{sfty_eqiup_use}{Safety equipment used by crash participant. Recorded as none, belt (seatbelt), unk (unkown), child (child restraint)}
  \item{airbag_deploy}{Records whether an airbag deployed (1) or not (0) in vehicle. Additional coding, 9, indicates airbag available on vehicle but deployment information is not given. Missing values (NA) indicate this information was not reported}
  \item{crash_month}{Month crash occured in}
  \item{day_of_month}{Day of the month crash occured on}
  \item{day_of_week}{Day of the week crash occured in coded as 1 (Sunday), 2 (Monday), 3 (Tuesday), 4 (Wednesday), 5 ( Thursday), 6 (Friday), 7 (Saturday) }
  \item{crash_year}{Year crash occured in}
  \item{longitude}{Longitude of crash location}
  \item{latitude}{Latitude of crash location}
  \item{partic_hit_run}{Whether or not a participant fled the scene of a crash. Coded 1 for "yes" and 0 for "no"}
  \item{rd_characteristic}{Type of road a collision occured on}
  \item{rd_type}{Indicates whether collision occured in the Portland highway system or Portland city street}
  \item{collision_type}{Indicates the type of collision. Levels are angl (vehicles collide at an angle on intersecting paths or roads), back (vehicle was backing up at the time of collision), fix (vehicle hit a fixed object), head (head-on collision caused by two cars attempting to occupy the same lane travelling opposite directions), ncol (non-collision, usually means the vehicle overturned), oth (other), park (at least one vehicle was parked at the time of collision), ped (vehicle collided with pedestrian), rear (vehicle rear-ended another vehicle), ss-m (sideswipe collision with vehicles are travelling opposite directions), ss-o (sideswipe collision with both vehicles travelling the same direction), and turn (vehicles collide while at least one is turning)}
  \item{speed_limit}{Indicates the posted speed limit at the location of crash}
  \item{crash_svrty}{Indicates the maximum severity of crash. Levels are fat (at least one person died), inj (at least one person was injured), and pdo (property damage only)}
  \item{weather_cond}{Indicates atmoshperic conditions at time of crash. Categories include unk (unknown), clr (clear), cld (cloudy), rain, slt (sleet or hail), fog, snow, and smok (smoke).}
  \item{rd_surface_cond}{Indicates the condition of travel lanes at the time of crash. This column may differ from weather_cond in scenarios that involve factors such as tunnels}
  \item{light_cond}{Indicates the amount of ambient light at time of crash. Categories are dark (night time with no streetlights), dlit (night time with streetlights present), dawn, day, dusk, and unk (unknown) .}
  \item{traf_device}{Indicates the type of traffic device that was most pertinent to crash cause. Categories are bus stopsign (bus stpsgn), emergency signal/flare (emr sgn/fl), flashing beacon-red (flashbcn-r), left turn green signal (l-grn-sig), l-turn ref (left turn refuge), median barrier (median bar), none, one way street (one-way), right turn green signal (r-grn-sig), ramp meter, special pedestrian signal (sp ped sig) stop sign, through green arrow or signal (thr-gn-sig), traffic signal (trf signal), unknown, warning, wigwag or flashing lights without drop arm gate (ww w/gate), yield}
  \item{bac_val}{Actual BAC test results. A value less than 80 represents the actual test result (80 means a BAC value of 0.80 or greater). A value of 86 means no test was administered, a value of 87 means a test was administered but results are unknown}
  \item{alcohol_use}{Indicates a participant's use of alcohol, according to police or that particular participant. A value of 9 indicates that it is unknown if a participant had been drinking}
  \item{drug_use}{Indicates use of any kind of drug, including prescription drugs, according to an officer, participant, or lab test. A value of 9 indicates unknown}
  \item{mj_use}{Indicates if a participant was impaired by cannabis according to lab tests or police. Because cannabis can be stored in the body for weeks, test results do not confirm impairment at the time of crash. Levels are 0 (negative), 1 (positive), 2 (suspected), 3 (not suspected), and 9 (unknown)}
  \item{speed_invlv}{Indicates if a driver involved in crash was driving faster than the speed limit, or too fast for conditions}
  \item{tot_vhcl_cnt}{Total number of vehicles involved in crash, excluding non-contact vehicles}
  \item{tot_fatal_cnt}{Total number of people killed as a result of crash}
  \item{tot_inj_cnt}{Total number of participants who suffered non-fatal injuries as a result of crash}
  \item{tot_ped_cnt}{Total number of pedestrians involved in crash}
  \item{tot_bicycle_cnt}{Total number of participants that were cyclists}
  \item{tot_per_inv_cnt}{Total count of all participants involved in crash}
}}
\source{
github
}
\usage{
portland_crashes_2018
}
\description{
A dataset containing information on all legally reportable car crashes and 
their participants in 2018 in Portland, OR.
}
\keyword{datasets}
